Bundle definition config mapping moved to another component.
