+++ api/src/graphql/graphql.controller.ts
@@ -1,6 +1,6 @@
 import {Controller, Post, Request, Response, Next} from '@nestjs/common';
 import {graphqlExpress} from 'apollo-server-express';
-import {GraphqlService} from './graphql.service';
+import {GraphqlService} from './service/graphql.service';
 
 @Controller()
 export class GraphqlController {
