+++ api/src/graphql/service/graphql.service.ts
@@ -1,20 +1,14 @@
-import {Component, HttpStatus, Inject} from '@nestjs/common';
+import {Component, Inject} from '@nestjs/common';
 import {GraphQLSchema} from 'graphql';
 import * as GraphQLJSON from 'graphql-type-json';
 import {makeExecutableSchema} from 'graphql-tools';
-import {ProjectRepository} from '../persistence/repository/project.repository';
-import {BuildDefinitionRepository} from '../persistence/repository/build-definition.repository';
-import {BuildInstanceRepository} from '../persistence/repository/build-instance.repository';
-import {ProjectTypeInterface} from './interfaces/project-type.interface';
-import {BuildDefinitionTypeInterface} from './interfaces/build-definition-type.interface';
-import {BuildInstanceTypeInterface} from './interfaces/build-instance-type.interface';
-import {BuildDefinitionConfigTypeInterface} from './interfaces/build-definition-config-type.interface';
-import {BuildDefinitionSourceTypeInterface} from './interfaces/build-definition-source-type.interface';
-import {BuildDefinitionSourceReferenceTypeInterface} from './interfaces/build-definition-source-reference-type.interface';
-import {BuildDefinitionProxiedPortTypeInterface} from './interfaces/build-definition-proxied-port-type.interface';
-import {BuildDefinitionSummaryItemTypeInterface} from './interfaces/build-definition-summary-item-type.interface';
-import {BuildDefinitionComposeFileTypeInterface} from './interfaces/build-definition-compose-file-type.interface';
-import {BuildDefinitionEnvironmentalVariableTypeInterface} from './interfaces/build-definition-environmental-variable-type.interface';
+import {ProjectRepository} from '../../persistence/repository/project.repository';
+import {BuildDefinitionRepository} from '../../persistence/repository/build-definition.repository';
+import {BuildInstanceRepository} from '../../persistence/repository/build-instance.repository';
+import {ProjectTypeInterface} from '../type/project-type.interface';
+import {BuildDefinitionTypeInterface} from '../type/build-definition-type.interface';
+import {BuildInstanceTypeInterface} from '../type/build-instance-type.interface';
+import {BundleDefinitionConfigMapper} from './bundle-definition-config-mapper.component';
 
 @Component()
 export class GraphqlService {
@@ -23,6 +17,7 @@ export class GraphqlService {
         private readonly projectRepository: ProjectRepository,
         private readonly buildDefinitionRepository: BuildDefinitionRepository,
         private readonly buildInstanceRepository: BuildInstanceRepository,
+        private readonly bundleDefinitionConfigMapper: BundleDefinitionConfigMapper,
     ) { }
 
     public get schema(): GraphQLSchema {
@@ -36,7 +31,6 @@ export class GraphqlService {
         return {
             JSON: GraphQLJSON,
             Query: {
-                hello: this.hello,
                 projects: this.projects,
                 buildDefinitions: this.buildDefinitions,
                 buildInstances: this.buildInstances,
@@ -58,12 +52,6 @@ export class GraphqlService {
             BuildDefinitionSummaryItem: {},
             BuildDefinitionEnvironmentalVariable: {},
             BuildDefinitionComposeFile: {},
-        };
-    }
-
-    public get hello(): () => Promise<string> {
-        return async (): Promise<string> => {
-            return 'World!';
         };
     }
 
@@ -93,7 +81,7 @@ export class GraphqlService {
                     id: buildDefinition._id,
                     name: buildDefinition.name,
                     projectId: buildDefinition.projectId,
-                    config: this.mapBundleDefinitionConfig(buildDefinition.config),
+                    config: this.bundleDefinitionConfigMapper.map(buildDefinition.config),
                 } as BuildDefinitionTypeInterface);
             }
 
@@ -173,109 +161,5 @@ export class GraphqlService {
 
             return data;
         };
-    }
-
-    protected mapBundleDefinitionConfig(config: any): BuildDefinitionConfigTypeInterface {
-        const mappedSources: BuildDefinitionSourceTypeInterface[] = [];
-        for (const sourceId of Object.keys(config.sources)) {
-            mappedSources.push(this.mapBundleDefinitionSource(sourceId, config.sources[sourceId]));
-        }
-
-        let mappedProxiedPorts: BuildDefinitionProxiedPortTypeInterface[] = [];
-        for (const containerName of Object.keys(config.exposedPorts)) {
-            mappedProxiedPorts = mappedProxiedPorts.concat(
-                this.mapBundleDefinitionProxiedPorts(containerName, config.exposedPorts[containerName]),
-            );
-        }
-
-        const mappedSummaryItems: BuildDefinitionSummaryItemTypeInterface[] = [];
-        for (const summaryItem of config.summaryItems) {
-            mappedSummaryItems.push(this.mapBundleDefinitionSummaryItem(summaryItem));
-        }
-
-        const mappedEnvironmentalVariables: BuildDefinitionEnvironmentalVariableTypeInterface[] = [];
-        if (config.environmentalVariables) {
-            for (const environmentalVariableName of Object.keys(config.environmentalVariables)) {
-                mappedEnvironmentalVariables.push(
-                    this.mapBundleDefinitionEnvironmentalVariable(
-                        environmentalVariableName,
-                        config.environmentalVariables[environmentalVariableName],
-                    ),
-                );
-            }
-        }
-
-        const mappedComposeFiles: BuildDefinitionComposeFileTypeInterface[] = [];
-        if (config.composeFiles) {
-            for (const composeFile of config.composeFiles) {
-                mappedComposeFiles.push(this.mapBundleDefinitionComposeFile(composeFile));
-            }
-        } else {
-            mappedComposeFiles.push(this.mapBundleDefinitionComposeFile(config.composeFile));
-        }
-
-        return {
-            sources: mappedSources,
-            proxiedPorts: mappedProxiedPorts,
-            summaryItems: mappedSummaryItems,
-            environmentalVariables: mappedEnvironmentalVariables,
-            composeFiles: mappedComposeFiles,
-        } as BuildDefinitionConfigTypeInterface;
-    }
-
-    protected mapBundleDefinitionSource(sourceId: string, source: any): BuildDefinitionSourceTypeInterface {
-        return {
-            id: sourceId,
-            type: source.type,
-            name: source.name,
-            reference: this.mapBundleDefinitionSourceReference(source.reference),
-        } as BuildDefinitionSourceTypeInterface;
-    }
-
-    protected mapBundleDefinitionSourceReference(reference: any): BuildDefinitionSourceReferenceTypeInterface {
-        return {
-            type: reference.type,
-            name: reference.name,
-        } as BuildDefinitionSourceReferenceTypeInterface;
-    }
-
-    protected mapBundleDefinitionProxiedPorts(containerName: string, proxiedPorts: any[]): BuildDefinitionProxiedPortTypeInterface[] {
-        const mappedProxiedPorts: BuildDefinitionProxiedPortTypeInterface[] = [];
-
-        for (const proxiedPort of proxiedPorts) {
-            mappedProxiedPorts.push(this.mapBundleDefinitionProxiedPort(containerName, proxiedPort));
-        }
-
-        return mappedProxiedPorts;
-    }
-
-    protected mapBundleDefinitionProxiedPort(containerName: string, proxiedPort: any): BuildDefinitionProxiedPortTypeInterface {
-        return {
-            id: proxiedPort.id,
-            containerName,
-            port: proxiedPort.port,
-            name: proxiedPort.name,
-        } as BuildDefinitionProxiedPortTypeInterface;
-    }
-
-    protected mapBundleDefinitionSummaryItem(summaryItem: any): BuildDefinitionSummaryItemTypeInterface {
-        return {
-            name: summaryItem.name,
-            text: summaryItem.value,
-        } as BuildDefinitionSummaryItemTypeInterface;
-    }
-
-    protected mapBundleDefinitionEnvironmentalVariable(name: string, value: string): BuildDefinitionEnvironmentalVariableTypeInterface {
-        return {
-            name,
-            value,
-        } as BuildDefinitionEnvironmentalVariableTypeInterface;
-    }
-
-    protected mapBundleDefinitionComposeFile(composeFile: any): BuildDefinitionComposeFileTypeInterface {
-        return {
-            sourceId: composeFile.sourceId,
-            relativePaths: composeFile.relativePaths || [composeFile.relativePath],
-        } as BuildDefinitionComposeFileTypeInterface;
     }
 }
