+++ api/src/graphql/type/build-definition-config-type.interface.ts
