+++ api/src/graphql/type/build-definition-environmental-variable-type.interface.ts
