+++ api/src/graphql/type/build-instance-type.interface.ts
